+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Fixes
 
+- Keep PropagationContext when cloning scope (#4518)
+
+## 8.40.1
+
+### Fixes
+
 - Session replay masking not working inside scroll view (#4498)
 
 ### Improvements